Clarify license and copyright notices
Included some changes to clarify recommendations for license notices and copyright notices. 
Note that this includes shifting the "Copyright the Zowe project..." line to be below the SPDX-License-Identifier. This is just to keep the license notice info together, so that the license header text and the SPDX ID aren't split by the copyright line.